[refactor][#48] test 관련 수정
