+++ src/test/java/kr/startoff/backend/interceptor/JwtInterceptorTest.java
@@ -9,7 +9,7 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.mock.web.MockHttpServletRequest;
 import org.springframework.mock.web.MockHttpServletResponse;
 
-import kr.startoff.backend.global.exception.custom.AccessTokenException;
+import kr.startoff.backend.domain.user.exception.AccessTokenException;
 import kr.startoff.backend.global.interceptor.JwtInterceptor;
 import kr.startoff.backend.global.security.jwt.CustomStatus;
